refactor: s3 pdf dropzone
