+++ frontend/src/features/lab-reports/dropzone/constants.ts
@@ -1,0 +1,2 @@
+export const MAX_FILE_SIZE_BYTES = 10 * 1024 * 1024; // 10 MB
+export const ACCEPTED_MIME = "application/pdf";
